pyproject.toml
[project]
name = "journal-check"
version = "0.1.0"
description = "Analyze psychology journal guidelines for pilot/feasibility study mentions"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "browser-use>=0.1.40",
    "playwright>=1.40.0",
    "instructor>=1.0.0",
    "google-generativeai>=0.8.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "rapidfuzz>=3.0.0",
    "lxml>=5.0.0",
    "langchain-google-genai>=4.2.0",
    "python-dotenv>=1.2.1",
    "google-genai>=1.61.0",
    "jsonref>=1.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
